requirements.txt
streamlit 
sentence-transformers 
pinecone-client